Pull out JSONEditor and add lint gutter markers
