+++ src/components/semantic/SemanticItem.tsx
@@ -1,20 +1,20 @@
 import React, { FunctionComponent, useRef, useState } from "react";
-import CodeMirror, { EditorView } from '@uiw/react-codemirror';
-import { linter } from '@codemirror/lint';
-import { json, jsonParseLinter } from '@codemirror/lang-json';
+import { Alert } from "reactstrap";
 
-import styles from "./styles.module.css";
+import { Content } from "../../isaac-data-types";
 
 import { ValuePresenterRef, ValueRef, ValueWrapper } from "./BaseValuePresenter";
-import { Content } from "../../isaac-data-types";
-import { getEntryType, PresenterProps } from "./registry";
-import { Alert, Button } from "reactstrap";
+import { getEntryType } from "./registry";
+import { JSONEditor } from "./JSONEditor";
+
+import styles from "./styles.module.css";
 
 export interface Shift {
     up: boolean;
     down: boolean;
     by: (amount: number) => void;
 }
+
 export interface SemanticItemProps {
     doc: Content;
     update: (newContent: Content) => void;
@@ -87,44 +87,6 @@ function SemanticItemInner(props: SemanticItemProps) {
 }
 
 const SemanticItemInnerMemo = React.memo(SemanticItemInner);
-
-const extensions = [json(), EditorView.lineWrapping, linter(jsonParseLinter())];
-
-const empty = Symbol("empty") as unknown as string;
-function JSONEditor({doc, update, close}: PresenterProps & {close: () => void}) {
-    const value = useRef(empty);
-    if (value.current === empty) {
-        value.current = JSON.stringify(doc, null, 2);
-    }
-    const [valid, setValid] = useState(true);
-    return <>
-        <CodeMirror
-            value={value.current}
-            maxHeight="calc(100vh - 120px)"
-            extensions={extensions}
-            onChange={(newValue) => {
-                value.current = newValue;
-                try {
-                    JSON.parse(newValue);
-                    setValid(true);
-                } catch (e) {
-                    console.error(e);
-                    setValid(false);
-                }
-            }}
-        />
-        <div className={styles.editButtons}>
-            <Button onClick={() => {
-                value.current = JSON.stringify(doc, null, 2);
-                close();
-            }}>Cancel</Button>
-            <Button color="primary" disabled={!valid} onClick={() => {
-                update(JSON.parse(value.current));
-                close();
-            }}>Save</Button>
-        </div>
-    </>;
-}
 
 export class SemanticItem extends React.Component<SemanticItemProps, { hasError: boolean; error: string; content?: string }> {
     constructor(props: SemanticItemProps) {
